Correct implementation of all possible payouts plus some bet fixing
Still need to do blackjack for human player

+++ Bot.java
@@ -4,7 +4,7 @@ public class Bot {
     //A new hand for the bot
   private Hand hand = new Hand();
 	private Random rand;
-	private int cpuWallet;
+	private double cpuWallet;
 	private int botBet;
 	private static int botWallet;
 
@@ -26,7 +26,7 @@ public class Bot {
 		}
 	}
 
-	public int getEachBotWallet(){
+	public double getEachBotWallet(){
 		return cpuWallet;
 	}
 
@@ -34,7 +34,7 @@ public class Bot {
 		cpuWallet = Player.getWallet();
 	}
 
-	public int getBotWallet(){
+	public double getBotWallet(){
 		return cpuWallet;
 	}
 
@@ -72,7 +72,7 @@ public class Bot {
 		return hand.total();
 	}
 
-	public void setBotWallet(int newWallet){
+	public void setBotWallet(double newWallet){
 		cpuWallet = newWallet;
 	}
 }
